+++ tests/integration/cli/test_tabular_format.py
@@ -38,7 +38,7 @@ def fxt_buddy_path(fxt_tabular_root):
 
 @pytest.fixture()
 def fxt_buddy_target():
-    yield ("breed_category", "pet_category")
+    yield {"input": "length(m)", "output": ["breed_category", "pet_category"]}
 
 
 @pytest.fixture()
@@ -76,12 +76,13 @@ class TabularIntegrationTest:
         dataset = request.getfixturevalue(fxt_dataset)
         path = request.getfixturevalue(fxt_path)
         target = request.getfixturevalue(fxt_target) if isinstance(fxt_target, str) else None
+        string_target = "input:length(m),output:breed_category,pet_category"
 
         with TestDir() as test_dir:
             run(helper_tc, "project", "create", "-o", test_dir)
             args = ["project", "import", "-p", test_dir, "-f", "tabular", path]
             if target:
-                args.extend(["--", "--target", ",".join(target)])
+                args.extend(["--", "--target", string_target])
             run(helper_tc, *args)
 
             export_dir = osp.join(test_dir, "export_dir")